refactor(run-once): extract pipeline support modules
