+++ src/cli/commands/run-once/pipeline.test.ts
@@ -1,19 +1,10 @@
 import test from "node:test";
 import assert from "node:assert/strict";
-import {
-  appendFile,
-  mkdir,
-  mkdtemp,
-  readFile,
-  writeFile,
-} from "node:fs/promises";
-import { tmpdir } from "node:os";
+import { mkdir, readFile, writeFile } from "node:fs/promises";
 import { join } from "node:path";
 import { DEFAULT_PATCHMILL_CONFIG } from "../../../config/defaults.ts";
 import { DEFAULT_PATCHMILL_POLICY } from "../../../policy/defaults.ts";
 import { createTriagePolicy } from "../../../policy/triage.ts";
-import { createPatchmillLabelCatalog } from "../../../policy/label-catalog.ts";
-import { createWorkflowApprovalPolicy } from "../../../workflow/approval-policy.ts";
 import {
   buildSkillPackMetadata,
   hashText,
@@ -24,30 +15,45 @@ import { runOneIssue } from "./pipeline.ts";
 import { JsonlProgressReporter } from "./progress.ts";
 import { assertNoLegacyProjectText } from "../../../../test-support/legacy-project-text.ts";
 import { formatPublishedArtifactComment } from "../../../workflow/artifacts/published-artifacts.ts";
-import type {
-  AgentIssueConfig,
-  AgentIssuePipelineResult,
-  AgentIssueProgressEvent,
-  CommandRunner,
-  CommandResult,
-  IssueSummary,
-} from "./types.ts";
-
-type Call = {
-  command: string;
-  args: string[];
-  cwd?: string;
-  env?: Record<string, string | undefined>;
-  onStdout?: (chunk: string) => void;
-  onStderr?: (chunk: string) => void;
-};
+import {
+  DEFAULT_LABEL_NAMES,
+  issue,
+  issueListPayload,
+  issueViewPayload,
+  labelListPayload,
+} from "../../../../test-support/run-once/issue-fixtures.ts";
+import {
+  appendPiSessionEntry,
+  assistantToolCall,
+  createMockRunner,
+  delay,
+  initializePiSession,
+  promptPath,
+  waitForCondition,
+  workflowPiCalls,
+  writePiSessionMessage,
+  type Call,
+} from "../../../../test-support/run-once/mock-runner.ts";
+import {
+  approvalPolicy,
+  blockedRecoveryRunner,
+  makeConfig,
+  runPlanApprovedImplementationScenario,
+  specAndPlanApprovalPolicy,
+  writeBlockedRecoveryRunState,
+} from "../../../../test-support/run-once/pipeline-fixtures.ts";
+import {
+  collectProgressEvents,
+  commentBody,
+  gitBaseContainmentFailure,
+  gitBaseContainmentResult,
+} from "../../../../test-support/run-once/assertions.ts";
+import type { CommandResult } from "./types.ts";
 
 const NOW = new Date("2026-05-09T12:00:00.000Z");
 const MINIMAL_PNG_BYTES = Buffer.from([
   0x89, 0x50, 0x4e, 0x47, 0x0d, 0x0a, 0x1a, 0x0a,
 ]);
-
-type MockRunner = CommandRunner & { calls: Call[] };
 
 function withRepo(args: string[], repoRoot: string): string[] {
   const separator = args.indexOf("--");
@@ -60,372 +66,12 @@ function withRepo(args: string[], repoRoot: string): string[] {
   ];
 }
 
-function commentBody(call: Call | undefined): string {
-  const separator = call?.args.indexOf("--") ?? -1;
-  return separator >= 0 ? (call?.args[separator + 1] ?? "") : "";
-}
-
-function issue(
-  number: number,
-  labels: string[],
-  title = `Issue ${number}`,
-): IssueSummary {
-  return {
-    number,
-    title,
-    body: `Body for issue ${number}`,
-    labels,
-    state: "open",
-    author: "rozanne",
-    updated: "2026-05-09T11:00:00Z",
-    comments: [
-      { author: { login: "ana" }, body: "Please keep this deterministic." },
-    ],
-  };
-}
-
-function teaIssuePayload(entry: IssueSummary) {
-  return {
-    index: entry.number,
-    title: entry.title,
-    body: entry.body,
-    state: entry.state,
-    labels: entry.labels.map((name) => ({ name })),
-    author: { login: entry.author },
-    updated: entry.updated,
-    comments: entry.comments,
-  };
-}
-
-function issueListPayload(issues: IssueSummary[]): string {
-  return JSON.stringify(issues.map(teaIssuePayload));
-}
-
-function issueViewPayload(issue: IssueSummary): string {
-  return JSON.stringify(teaIssuePayload(issue));
-}
-
 const LANDING_SKILLS = {
   ...DEFAULT_PATCHMILL_CONFIG.skills,
   landing: "project-landing",
 };
 
 const cleanupHook = "./scripts/cleanup.sh";
-
-const DEFAULT_LABEL_NAMES = [
-  "agent-ready",
-  "needs-info",
-  "agent-unsuitable",
-  "in-progress",
-  "agent-done",
-  "bug",
-  "enhancement",
-  "docs",
-  "chore",
-  "test",
-  "priority:low",
-  "priority:medium",
-  "priority:high",
-  "priority:critical",
-  "spec-review",
-  "spec-approved",
-  "plan-review",
-  "plan-approved",
-] as const;
-
-function labelListPayload(
-  labels: readonly string[] = DEFAULT_LABEL_NAMES,
-): string {
-  return JSON.stringify(labels.map((name) => ({ name })));
-}
-
-function collectProgressEvents(): {
-  events: AgentIssueProgressEvent[];
-  progress: { event: (event: AgentIssueProgressEvent) => void };
-} {
-  const events: AgentIssueProgressEvent[] = [];
-  return {
-    events,
-    progress: {
-      event: (event) => {
-        events.push(event);
-      },
-    },
-  };
-}
-
-function normalizeRecordedPiCall(call: Call): Call {
-  if (
-    call.command === process.execPath &&
-    /@earendil-works[/\\]pi-coding-agent[/\\]dist[/\\]cli\.js$/.test(
-      call.args[0] ?? "",
-    )
-  ) {
-    return { ...call, command: "pi", args: call.args.slice(1) };
-  }
-  return call;
-}
-
-function gitBaseContainmentResult(call: Call): CommandResult | undefined {
-  if (call.command !== "git") return undefined;
-  if (call.args[0] === "rev-parse" && call.args[1] === "--verify") {
-    return { code: 0, stdout: "commit-sha\n", stderr: "" };
-  }
-  if (
-    call.args[0] === "log" &&
-    call.args[1] === "--oneline" &&
-    call.args[2]?.startsWith("refs/remotes/")
-  ) {
-    return { code: 0, stdout: "", stderr: "" };
-  }
-  return undefined;
-}
-
-function gitBaseContainmentFailure(call: Call): CommandResult | undefined {
-  if (call.command !== "git") return undefined;
-  if (call.args[0] === "rev-parse" && call.args[1] === "--verify") {
-    return { code: 0, stdout: "commit-sha\n", stderr: "" };
-  }
-  if (
-    call.args[0] === "log" &&
-    call.args[1] === "--oneline" &&
-    call.args[2]?.startsWith("refs/remotes/")
-  ) {
-    return {
-      code: 0,
-      stdout: "abc1234 chore: initialize Patchmill\n",
-      stderr: "",
-    };
-  }
-  return undefined;
-}
-
-function workflowPiCalls(calls: Call[]): Call[] {
-  return calls.filter((call) => call.command === "pi");
-}
-
-function createMockRunner(
-  handler: (call: Call) => Promise<CommandResult> | CommandResult,
-): MockRunner {
-  const calls: Call[] = [];
-  return {
-    calls,
-    async run(command, args, options = {}) {
-      const call = normalizeRecordedPiCall({
-        command,
-        args: [...args],
-        cwd: options.cwd,
-        ...(options.env ? { env: options.env } : {}),
-        onStdout: options.onStdout,
-        onStderr: options.onStderr,
-      });
-      calls.push(call);
-      if (call.command === "pi") {
-        try {
-          return await normalizePiResult(call, await handler(call));
-        } catch (error) {
-          const fallback = await fallbackPiResultForError(call);
-          if (fallback) return fallback;
-          throw error;
-        }
-      }
-      const baseContainment = gitBaseContainmentResult(call);
-      if (baseContainment) {
-        try {
-          const result = await handler(call);
-          if (result.stdout.includes("abc1234 chore: initialize Patchmill")) {
-            return result;
-          }
-        } catch {
-          // Most existing tests predate this preflight and throw on the new git
-          // commands. Treat those as a clean base unless a test explicitly
-          // returns the unsafe marker above.
-        }
-        return baseContainment;
-      }
-      return await handler(call);
-    },
-  };
-}
-
-function promptPath(args: string[]): string {
-  const promptArg = args.find((arg) => arg.startsWith("@"));
-  assert.ok(promptArg, `expected prompt path in ${args.join(" ")}`);
-  return promptArg.slice(1);
-}
-
-function jsonStatus(stdout: string): string | undefined {
-  return stdout.match(/"status"\s*:\s*"([^"]+)"/)?.[1];
-}
-
-function promptJsonPath(prompt: string, key: "specPath" | "planPath"): string {
-  const match = prompt.match(new RegExp(`"${key}"\\s*:\\s*"([^"]+)"`));
-  assert.ok(match?.[1], `expected ${key} in prompt`);
-  return match[1];
-}
-
-function defaultWorkflowPromptResult(
-  prompt: string,
-): CommandResult | undefined {
-  if (/Create a design spec/.test(prompt)) {
-    return {
-      code: 0,
-      stdout: JSON.stringify({
-        status: "spec-created",
-        specPath: promptJsonPath(prompt, "specPath"),
-        commit: "spec123",
-      }),
-      stderr: "",
-    };
-  }
-
-  if (/Create an implementation plan/.test(prompt)) {
-    return {
-      code: 0,
-      stdout: JSON.stringify({
-        status: "plan-created",
-        planPath: promptJsonPath(prompt, "planPath"),
-        commit: "abc123",
-      }),
-      stderr: "",
-    };
-  }
-
-  return undefined;
-}
-
-async function normalizePiResult(
-  call: Call,
-  result: CommandResult,
-): Promise<CommandResult> {
-  const prompt = await readFile(promptPath(call.args), "utf8");
-  const fallback = defaultWorkflowPromptResult(prompt);
-  if (!fallback) return result;
-
-  if (result.code !== 0) return result;
-
-  const status = jsonStatus(result.stdout);
-  if (status === "blocked") return result;
-  if (/Create a design spec/.test(prompt)) {
-    return status === "spec-created" ? result : fallback;
-  }
-  if (/Create an implementation plan/.test(prompt)) {
-    return status === "plan-created" ? result : fallback;
-  }
-
-  return result;
-}
-
-async function fallbackPiResultForError(
-  call: Call,
-): Promise<CommandResult | undefined> {
-  const prompt = await readFile(promptPath(call.args), "utf8");
-  return defaultWorkflowPromptResult(prompt);
-}
-
-async function writePiSessionMessage(
-  call: Call,
-  text: string,
-  usage?: { input: number; output: number; totalTokens: number },
-): Promise<void> {
-  const sessionDirIndex = call.args.indexOf("--session-dir");
-  assert.ok(
-    sessionDirIndex >= 0,
-    `expected --session-dir in ${call.args.join(" ")}`,
-  );
-  const sessionDir = call.args[sessionDirIndex + 1];
-  assert.ok(sessionDir);
-  const sessionSubdir = join(sessionDir, "--repo--");
-  await mkdir(sessionSubdir, { recursive: true });
-  await writeFile(
-    join(sessionSubdir, "session.jsonl"),
-    [
-      JSON.stringify({
-        type: "session",
-        version: 3,
-        id: "session-1",
-        cwd: call.cwd,
-      }),
-      JSON.stringify({
-        type: "message",
-        id: "assistant-1",
-        parentId: null,
-        timestamp: "2026-05-09T12:00:00.000Z",
-        message: {
-          role: "assistant",
-          content: [{ type: "text", text }],
-          usage,
-        },
-      }),
-    ].join("\n") + "\n",
-    "utf8",
-  );
-}
-
-async function piSessionPath(call: Call): Promise<string> {
-  const sessionDirIndex = call.args.indexOf("--session-dir");
-  assert.ok(
-    sessionDirIndex >= 0,
-    `expected --session-dir in ${call.args.join(" ")}`,
-  );
-  const sessionDir = call.args[sessionDirIndex + 1];
-  assert.ok(sessionDir);
-  const sessionSubdir = join(sessionDir, "--repo--");
-  await mkdir(sessionSubdir, { recursive: true });
-  return join(sessionSubdir, "session.jsonl");
-}
-
-async function appendPiSessionEntry(call: Call, entry: unknown): Promise<void> {
-  await appendFile(
-    await piSessionPath(call),
-    `${JSON.stringify(entry)}\n`,
-    "utf8",
-  );
-}
-
-async function initializePiSession(call: Call): Promise<void> {
-  await writeFile(
-    await piSessionPath(call),
-    `${JSON.stringify({ type: "session", version: 3, id: "session-1", cwd: call.cwd })}\n`,
-    "utf8",
-  );
-}
-
-function assistantToolCall(
-  toolCallId: string,
-  toolName: string,
-  args: Record<string, unknown>,
-): unknown {
-  return {
-    type: "message",
-    id: `assistant-${toolCallId}`,
-    parentId: null,
-    timestamp: "2026-05-09T12:00:00.000Z",
-    message: {
-      role: "assistant",
-      content: [
-        { type: "toolCall", id: toolCallId, name: toolName, arguments: args },
-      ],
-    },
-  };
-}
-
-function delay(ms: number): Promise<void> {
-  return new Promise((resolve) => setTimeout(resolve, ms));
-}
-
-async function waitForCondition(
-  condition: () => boolean,
-  failureMessage: () => string,
-  timeoutMs = 1_000,
-): Promise<void> {
-  const deadline = Date.now() + timeoutMs;
-  while (Date.now() < deadline) {
-    if (condition()) return;
-    await delay(5);
-  }
-  assert.ok(condition(), failureMessage());
-}
 
 async function writeTodo(
   repoRoot: string,
@@ -440,203 +86,6 @@ async function writeTodo(
     `${JSON.stringify({ id, title, status })}\n\nbody\n`,
     "utf8",
   );
-}
-
-function specAndPlanApprovalPolicy() {
-  return approvalPolicy({ specRequired: true, planRequired: true });
-}
-
-function approvalPolicy(
-  overrides: {
-    specRequired?: boolean;
-    specApprovedLabel?: string;
-    planRequired?: boolean;
-    planReviewLabel?: string;
-    planApprovedLabel?: string;
-  } = {},
-) {
-  return createWorkflowApprovalPolicy({
-    ...DEFAULT_PATCHMILL_CONFIG.workflow,
-    specApproval: {
-      ...DEFAULT_PATCHMILL_CONFIG.workflow.specApproval,
-      required:
-        overrides.specRequired ??
-        DEFAULT_PATCHMILL_CONFIG.workflow.specApproval.required,
-      approvedLabel:
-        overrides.specApprovedLabel ??
-        DEFAULT_PATCHMILL_CONFIG.workflow.specApproval.approvedLabel,
-    },
-    planApproval: {
-      ...DEFAULT_PATCHMILL_CONFIG.workflow.planApproval,
-      required:
-        overrides.planRequired ??
-        DEFAULT_PATCHMILL_CONFIG.workflow.planApproval.required,
-      reviewLabel:
-        overrides.planReviewLabel ??
-        DEFAULT_PATCHMILL_CONFIG.workflow.planApproval.reviewLabel,
-      approvedLabel:
-        overrides.planApprovedLabel ??
-        DEFAULT_PATCHMILL_CONFIG.workflow.planApproval.approvedLabel,
-    },
-  });
-}
-
-async function makeConfig(
-  overrides: Partial<AgentIssueConfig> = {},
-): Promise<AgentIssueConfig> {
-  const repoRoot = await mkdtemp(join(tmpdir(), "patchmill-issue-pipeline-"));
-  const specsDir = join(repoRoot, "docs", "specs");
-  const plansDir = join(repoRoot, "docs", "plans");
-  const runStateDir = join(repoRoot, ".patchmill", "runs");
-  await mkdir(specsDir, { recursive: true });
-  await mkdir(plansDir, { recursive: true });
-  const labelCatalog = createPatchmillLabelCatalog({
-    ...DEFAULT_PATCHMILL_CONFIG,
-    labels: overrides.triagePolicy?.labels ?? DEFAULT_PATCHMILL_CONFIG.labels,
-    triage: overrides.triagePolicy
-      ? { stateMap: overrides.triagePolicy.stateMap }
-      : DEFAULT_PATCHMILL_CONFIG.triage,
-  });
-
-  return {
-    repoRoot,
-    dryRun: true,
-    execute: false,
-    planOnly: false,
-    host: { provider: "forgejo-tea", login: "" },
-    specsDir,
-    plansDir,
-    runStateDir,
-    worktreeDir: join(repoRoot, ".worktrees"),
-    cleanStatusIgnorePrefixes: [".patchmill/runs/", ".patchmill/triage-runs/"],
-    projectPolicy: DEFAULT_PATCHMILL_POLICY,
-    readyLabel: "agent-ready",
-    issueLimit: 1,
-    labelCatalog,
-    approvalPolicy: createWorkflowApprovalPolicy(
-      DEFAULT_PATCHMILL_CONFIG.workflow,
-    ),
-    baseBranch: "main",
-    baseRef: "HEAD",
-    remote: "origin",
-    branchPrefix: "agent/issue-",
-    worktreePrefix: "patchmill-issue-",
-    slugLength: 48,
-    allowDirectLand: true,
-    skills: { ...DEFAULT_PATCHMILL_CONFIG.skills },
-    ...overrides,
-  };
-}
-
-type PlanApprovedImplementationScenario = {
-  issueNumber: number;
-  title: string;
-  issueLabels?: string[];
-  planPath?: string;
-  configOverrides?: Partial<AgentIssueConfig>;
-  onPi?: (input: {
-    call: Call;
-    prompt: string;
-    config: AgentIssueConfig;
-    piPrompts: string[];
-  }) => CommandResult | Promise<CommandResult>;
-};
-
-async function runPlanApprovedImplementationScenario(
-  scenario: PlanApprovedImplementationScenario,
-): Promise<{
-  config: AgentIssueConfig;
-  runner: MockRunner;
-  result: AgentIssuePipelineResult;
-  piPrompts: string[];
-  selected: IssueSummary;
-}> {
-  const config = await makeConfig({
-    dryRun: false,
-    execute: true,
-    ...scenario.configOverrides,
-  });
-  const planPath =
-    scenario.planPath ??
-    `docs/plans/2026-05-14-issue-${scenario.issueNumber}-scenario.md`;
-  await writeFile(join(config.repoRoot, planPath), "# plan\n", "utf8");
-  const selected = issue(
-    scenario.issueNumber,
-    scenario.issueLabels ?? ["plan-approved"],
-    scenario.title,
-  );
-  const piPrompts: string[] = [];
-  const runner = createMockRunner(async (call) => {
-    if (
-      call.command === "tea" &&
-      call.args[0] === "issues" &&
-      call.args[1] === "list"
-    ) {
-      const page = call.args[call.args.indexOf("--page") + 1];
-      return {
-        code: 0,
-        stdout: page === "1" ? issueListPayload([selected]) : "[]",
-        stderr: "",
-      };
-    }
-    if (call.command === "git" && call.args[0] === "status") {
-      return { code: 0, stdout: "", stderr: "" };
-    }
-    if (
-      call.command === "git" &&
-      call.args[0] === "worktree" &&
-      call.args[1] === "list"
-    ) {
-      return { code: 0, stdout: "", stderr: "" };
-    }
-    if (call.command === "git" && call.args[0] === "show-ref") {
-      return { code: 1, stdout: "", stderr: "" };
-    }
-    if (
-      call.command === "git" &&
-      call.args[0] === "worktree" &&
-      call.args[1] === "add"
-    ) {
-      return { code: 0, stdout: "", stderr: "" };
-    }
-    if (
-      call.command === "tea" &&
-      call.args[0] === "labels" &&
-      call.args[1] === "list"
-    ) {
-      return { code: 0, stdout: labelListPayload(), stderr: "" };
-    }
-    if (call.command === "tea" && call.args[0] === "comment") {
-      return { code: 0, stdout: "", stderr: "" };
-    }
-    if (call.command === "tea" && call.args[0] === "issues") {
-      return { code: 0, stdout: "", stderr: "" };
-    }
-    if (call.command === "pi") {
-      const prompt = await readFile(promptPath(call.args), "utf8");
-      piPrompts.push(prompt);
-      return scenario.onPi
-        ? await scenario.onPi({ call, prompt, config, piPrompts })
-        : {
-            code: 0,
-            stdout: JSON.stringify({
-              status: "pr-created",
-              prUrl: `https://forgejo.example/pr/${scenario.issueNumber}`,
-              branch: `agent/issue-${scenario.issueNumber}-implementation`,
-              commits: ["123abc"],
-              validation: ["npm test"],
-              reviewSummary: "reviewed",
-            }),
-            stderr: "",
-          };
-    }
-    throw new Error(
-      `unexpected command: ${call.command} ${call.args.join(" ")}`,
-    );
-  });
-
-  const result = await runOneIssue(runner, config, { now: NOW });
-  return { config, runner, result, piPrompts, selected };
 }
 
 test("runOneIssue dry-run lists open issues and returns the selected agent-ready issue without mutations", async () => {
@@ -3571,218 +3020,6 @@ test("runOneIssue reuses existing implementation worktree on resume", async () =
   );
   assert.ok(runner.calls.find((call) => call.command === "pi"));
 });
-
-async function writeBlockedRecoveryRunState(
-  config: AgentIssueConfig,
-  overrides: Parameters<typeof writeRunState>[1] = {
-    issueNumber: 45,
-    status: "blocked",
-  },
-  options: {
-    createWorktreePath?: boolean;
-    writePlanInPrimaryRepo?: boolean;
-    writeSpecInPrimaryRepo?: boolean;
-    writePlanInWorktree?: boolean;
-    writeSpecInWorktree?: boolean;
-  } = {},
-): Promise<void> {
-  const planPath =
-    overrides.planPath ??
-    "docs/plans/2026-06-20-issue-45-recover-blocked-run.md";
-  const specPath =
-    overrides.specPath ??
-    "docs/specs/2026-06-20-issue-45-recover-blocked-run.md";
-  const worktreePath =
-    overrides.worktreePath ??
-    ".worktrees/patchmill-issue-45-recover-blocked-run";
-  const worktreeRoot = join(config.repoRoot, worktreePath);
-
-  if (options.writePlanInPrimaryRepo !== false) {
-    await writeFile(join(config.repoRoot, planPath), "# plan\n", "utf8");
-  }
-  if (options.writeSpecInPrimaryRepo === true) {
-    await writeFile(join(config.repoRoot, specPath), "# spec\n", "utf8");
-  }
-
-  if (options.createWorktreePath !== false) {
-    await mkdir(worktreeRoot, { recursive: true });
-  }
-  if (options.writePlanInWorktree) {
-    await mkdir(join(worktreeRoot, "docs", "plans"), { recursive: true });
-    await writeFile(join(worktreeRoot, planPath), "# worktree plan\n", "utf8");
-  }
-  if (options.writeSpecInWorktree) {
-    await mkdir(join(worktreeRoot, "docs", "specs"), { recursive: true });
-    await writeFile(join(worktreeRoot, specPath), "# worktree spec\n", "utf8");
-  }
-
-  await writeRunState(
-    config.runStateDir,
-    {
-      issueNumber: 45,
-      title: "Recover blocked run",
-      status: "blocked",
-      specPath,
-      specCommit: "spec123",
-      planPath,
-      planCommit: "plan123",
-      branch: "agent/issue-45-recover-blocked-run",
-      worktreePath,
-      commits: ["abc123", "def456"],
-      validation: ["formatting passed", "verification environment unavailable"],
-      failureCommentKeys: ["blocked:verification"],
-      lastError: "Required verification environment is unavailable.",
-      checkpoints: {
-        claimed: true,
-        startedCommentPosted: true,
-        specPathResolved: true,
-        planPathResolved: true,
-        worktreeReady: true,
-      },
-      ...overrides,
-    },
-    NOW.toISOString(),
-  );
-}
-
-function blockedRecoveryRunner(
-  config: AgentIssueConfig,
-  options: {
-    selectedLabels?: string[];
-    branchExists?: boolean;
-    worktreeRegistered?: boolean;
-    dirtyStatus?: string;
-    merged?: boolean;
-    revList?: string;
-    log?: string;
-    onPi?: (prompt: string) => CommandResult;
-    selectedComments?: IssueSummary["comments"];
-  } = {},
-): MockRunner {
-  return createMockRunner(async (call) => {
-    if (
-      call.command === "tea" &&
-      call.args[0] === "issues" &&
-      call.args[1] === "list"
-    ) {
-      const page = call.args[call.args.indexOf("--page") + 1];
-      return {
-        code: 0,
-        stdout:
-          page === "1"
-            ? issueListPayload([
-                {
-                  ...issue(
-                    45,
-                    options.selectedLabels ?? ["needs-info"],
-                    "Recover blocked run",
-                  ),
-                  ...(options.selectedComments
-                    ? { comments: options.selectedComments }
-                    : {}),
-                },
-              ])
-            : "[]",
-        stderr: "",
-      };
-    }
-    if (call.command === "git" && call.args[0] === "show-ref") {
-      return {
-        code: options.branchExists === false ? 1 : 0,
-        stdout: "",
-        stderr: "",
-      };
-    }
-    if (
-      call.command === "git" &&
-      call.args[0] === "worktree" &&
-      call.args[1] === "list"
-    ) {
-      return {
-        code: 0,
-        stdout:
-          options.worktreeRegistered === false
-            ? ""
-            : `worktree ${join(config.repoRoot, ".worktrees/patchmill-issue-45-recover-blocked-run")}\n`,
-        stderr: "",
-      };
-    }
-    if (
-      call.command === "git" &&
-      call.args[0] === "-C" &&
-      call.args[2] === "status"
-    ) {
-      return { code: 0, stdout: options.dirtyStatus ?? "", stderr: "" };
-    }
-    if (
-      call.command === "git" &&
-      call.args[0] === "-C" &&
-      call.args[2] === "branch"
-    ) {
-      return {
-        code: 0,
-        stdout: "agent/issue-45-recover-blocked-run\n",
-        stderr: "",
-      };
-    }
-    if (call.command === "git" && call.args[0] === "merge-base") {
-      return { code: options.merged ? 0 : 1, stdout: "", stderr: "" };
-    }
-    if (call.command === "git" && call.args[0] === "rev-list") {
-      return { code: 0, stdout: options.revList ?? "0\t2\n", stderr: "" };
-    }
-    if (call.command === "git" && call.args[0] === "status") {
-      return { code: 0, stdout: "", stderr: "" };
-    }
-    if (call.command === "git" && call.args[0] === "log") {
-      return {
-        code: 0,
-        stdout:
-          options.log ?? "def456 add verification\nabc123 implement feature\n",
-        stderr: "",
-      };
-    }
-    if (call.command === "tea" && call.args[0] === "logins")
-      return {
-        code: 0,
-        stdout: JSON.stringify([
-          { name: "default", user: "patchmill-bot", default: true },
-        ]),
-        stderr: "",
-      };
-    if (
-      call.command === "tea" &&
-      call.args[0] === "labels" &&
-      call.args[1] === "list"
-    )
-      return { code: 0, stdout: labelListPayload(), stderr: "" };
-    if (
-      call.command === "tea" &&
-      (call.args[0] === "issues" || call.args[0] === "comment")
-    )
-      return { code: 0, stdout: "", stderr: "" };
-    if (call.command === "pi") {
-      const prompt = await readFile(promptPath(call.args), "utf8");
-      return options.onPi
-        ? options.onPi(prompt)
-        : {
-            code: 0,
-            stdout: JSON.stringify({
-              status: "pr-created",
-              prUrl: "https://forgejo/pr/45",
-              branch: "agent/issue-45-recover-blocked-run",
-              commits: ["abc123", "def456", "789abc"],
-              validation: ["npm test passed"],
-              reviewSummary: "reviewed",
-            }),
-            stderr: "",
-          };
-    }
-    throw new Error(
-      `unexpected command: ${call.command} ${call.args.join(" ")}`,
-    );
-  });
-}
 
 test("runOneIssue resumes clean blocked implementation workspace after external prerequisite is fixed", async () => {
   const config = await makeConfig({
