+++ src/cli/commands/run-once/pipeline.ts
@@ -1,15 +1,10 @@
-import { isAbsolute, join, relative } from "node:path";
+import { join } from "node:path";
 import { runCleanupHookScript } from "../../../pi/hooks.ts";
 import { localPiAgentDir } from "../init/pi-agent-settings.ts";
-import {
-  buildIssueBranchName,
-  buildIssueWorktreePath,
-} from "../../../git/worktree-strategy.ts";
-import type { GitWorktreeStrategyConfig } from "../../../git/types.ts";
+
 import { createIssueHostProvider } from "../../../host/factory.ts";
-import type { IssueHostProvider } from "../../../host/types.ts";
 import { skillInvocationPaths } from "../../../workflow/skills.ts";
-import { DEFAULT_TRIAGE_POLICY, planLabelChange } from "../triage/labels.ts";
+import { planLabelChange } from "../triage/labels.ts";
 import { materializeIssueArtifactSources } from "./artifact-source-materialization.ts";
 import { runArtifactSourceStage } from "./artifact-source-stage.ts";
 import type { ResolvedIssueArtifactSources } from "./artifact-sources.ts";
@@ -17,7 +12,6 @@ import { ensureAutomationLabel } from "./automation-labels.ts";
 import {
   assertCleanWorktree,
   assertIssueBaseContainedInPrBase,
-  cleanStatusIgnoredPaths as buildCleanStatusIgnoredPaths,
   cleanupIssueWorkspace,
   ensureIssueWorktree,
   type IssueWorktreeResult,
@@ -41,7 +35,6 @@ import {
   ApprovalRequiredError,
   cleanupLabelsForImplementation,
   resolveWorkflowState,
-  type RunOnceWorkflowState,
 } from "./workflow-state.ts";
 import {
   isResumableRunState,
@@ -54,20 +47,38 @@ import {
   hasBlockedRunRecoveryState,
   inspectBlockedRunRecovery,
 } from "./recovery.ts";
-import { selectIssue, selectIssueWithDiagnostics } from "./selection.ts";
+import { selectIssueWithDiagnostics } from "./selection.ts";
+import { emitSimpleStep, progress, withLogPath } from "./pipeline-progress.ts";
+import {
+  AgentIssueSafetyError,
+  assertDirectLandAllowed,
+  effectiveCheckpoints,
+  lifecycleLabels,
+  nextLabels,
+  successfulImplementationFromState,
+  workflowTransition,
+} from "./pipeline-lifecycle.ts";
+import { handoffComment, startedComment } from "./pipeline-comments.ts";
+import {
+  cleanStatusIgnoredPaths,
+  configuredWorktreeStrategy,
+  expectedIssueWorkspace,
+  resumePlanningArtifactPolicy,
+} from "./pipeline-workspace.ts";
+import {
+  emitSelectionDiagnostics,
+  loadSelectionIssues,
+  selectResumableIssue,
+} from "./pipeline-selection.ts";
+import { blockIssue, unexpectedFailure } from "./pipeline-failures.ts";
 import { validateVisualEvidenceReferences } from "./visual-evidence.ts";
 import type { AgentIssueProgressEvent, ProgressReporter } from "./progress.ts";
 import type {
-  AgentIssueBlockedResult,
-  AgentIssueBlockerQuestion,
   AgentIssueConfig,
   AgentIssueDevelopmentEnvironmentHandoff,
   AgentIssuePiResult,
   AgentIssuePipelineResult,
-  AgentIssueVisualEvidence,
-  AgentIssueRunCheckpoints,
   CommandRunner,
-  IssueSelectionRejection,
   IssueSummary,
 } from "./types.ts";
 
@@ -79,744 +90,6 @@ export type RunOneIssueOptions = {
   verbosePiOutput?: boolean;
   heartbeatMs?: number;
 };
-
-async function progress(
-  options: RunOneIssueOptions,
-  level: AgentIssueProgressEvent["level"],
-  stage: string,
-  message: string,
-  extras: Partial<
-    Pick<
-      AgentIssueProgressEvent,
-      "issueNumber" | "elapsedSeconds" | "data" | "consoleMessage"
-    >
-  > = {},
-): Promise<void> {
-  await options.progress?.event({
-    time: (options.now ?? new Date()).toISOString(),
-    level,
-    stage,
-    message,
-    ...extras,
-  });
-}
-
-async function emitSimpleStep(
-  options: RunOneIssueOptions,
-  issueNumber: number,
-  label: string,
-): Promise<void> {
-  const time = new Date().toISOString();
-  await options.progress?.event({
-    time,
-    level: "info",
-    stage: "step",
-    message: label,
-    issueNumber,
-    step: { type: "step-start", label },
-  });
-  await options.progress?.event({
-    time: new Date().toISOString(),
-    level: "info",
-    stage: "step",
-    message: label,
-    issueNumber,
-    step: { type: "step-complete", label },
-  });
-}
-
-function withLogPath<T extends AgentIssuePipelineResult>(
-  result: T,
-  options: RunOneIssueOptions,
-): T {
-  return options.logPath ? { ...result, logPath: options.logPath } : result;
-}
-
-function cleanStatusIgnoredPaths(
-  config: Pick<
-    AgentIssueConfig,
-    "runStateDir" | "cleanStatusIgnorePrefixes" | "projectPolicy"
-  >,
-  options: Pick<RunOneIssueOptions, "logPath">,
-): string[] {
-  return buildCleanStatusIgnoredPaths({
-    cleanStatusIgnorePrefixes: config.cleanStatusIgnorePrefixes,
-    todoRoot: config.projectPolicy.pi.taskContract.todoRoot,
-    runStateDir: config.runStateDir,
-    additionalPaths: options.logPath ? [options.logPath] : [],
-  });
-}
-
-function configuredWorktreeDir(
-  config: Pick<AgentIssueConfig, "repoRoot" | "worktreeDir">,
-): string {
-  return relative(config.repoRoot, config.worktreeDir) || ".";
-}
-
-function configuredPathRelativeToRepo(repoRoot: string, path: string): string {
-  return isAbsolute(path) ? relative(repoRoot, path) : path;
-}
-
-function mirrorConfiguredPathInWorktree(
-  repoRoot: string,
-  worktreeRoot: string,
-  path: string,
-): string {
-  return join(worktreeRoot, configuredPathRelativeToRepo(repoRoot, path));
-}
-
-function resumePlanningArtifactPolicy(input: {
-  config: Pick<AgentIssueConfig, "repoRoot" | "specsDir" | "plansDir">;
-  worktreePath: string;
-  existingState: NonNullable<Awaited<ReturnType<typeof readRunState>>>;
-  resolvedArtifacts: ResolvedIssueArtifactSources;
-}): PlanningArtifactPolicy {
-  const worktreeRoot = join(input.config.repoRoot, input.worktreePath);
-  return {
-    kind: "implementation-resume",
-    primary: {
-      repoRoot: worktreeRoot,
-      specsDir: mirrorConfiguredPathInWorktree(
-        input.config.repoRoot,
-        worktreeRoot,
-        input.config.specsDir,
-      ),
-      plansDir: mirrorConfiguredPathInWorktree(
-        input.config.repoRoot,
-        worktreeRoot,
-        input.config.plansDir,
-      ),
-      source: "resume-worktree",
-    },
-    fallbacks: [
-      {
-        repoRoot: input.config.repoRoot,
-        specsDir: input.config.specsDir,
-        plansDir: input.config.plansDir,
-        source: "primary-repo",
-      },
-    ],
-    saved: {
-      specPath: input.existingState.specPath,
-      specCommit: input.existingState.specCommit,
-      planPath: input.existingState.planPath,
-      planCommit: input.existingState.planCommit,
-      specCreated: input.existingState.checkpoints?.specCreated,
-      planCreated: input.existingState.checkpoints?.planCreated,
-    },
-    explicit: input.resolvedArtifacts,
-  };
-}
-
-function configuredWorktreeStrategy(
-  config: Pick<
-    AgentIssueConfig,
-    keyof GitWorktreeStrategyConfig | "repoRoot" | "worktreeDir"
-  >,
-): GitWorktreeStrategyConfig {
-  return {
-    baseBranch: config.baseBranch,
-    baseRef: config.baseRef,
-    remote: config.remote,
-    branchPrefix: config.branchPrefix,
-    worktreeDir: configuredWorktreeDir(config),
-    worktreePrefix: config.worktreePrefix,
-    slugLength: config.slugLength,
-    allowDirectLand: config.allowDirectLand,
-  };
-}
-
-function expectedIssueWorkspace(
-  issueNumber: number,
-  title: string,
-  strategy: GitWorktreeStrategyConfig,
-): {
-  branch: string;
-  worktreePath: string;
-} {
-  return {
-    branch: buildIssueBranchName(issueNumber, title, strategy),
-    worktreePath: buildIssueWorktreePath(issueNumber, title, strategy),
-  };
-}
-
-function nextLabels(
-  labels: string[],
-  remove: string[],
-  add: string[],
-): string[] {
-  const removed = new Set(remove);
-  const kept = labels.filter((label) => !removed.has(label));
-  return [...kept, ...add.filter((label) => !kept.includes(label))];
-}
-
-function workflowTransition(
-  state: RunOnceWorkflowState,
-  config: Pick<AgentIssueConfig, "approvalPolicy">,
-): string {
-  if (state.kind === "plan-approved") return "plan-approved -> agent-done";
-  if (state.kind === "spec-approved") {
-    return config.approvalPolicy.planApproval.required
-      ? "spec-approved -> plan-review"
-      : "spec-approved -> agent-done";
-  }
-  if (state.kind === "agent-ready") {
-    if (config.approvalPolicy.specApproval.required) {
-      return "agent-ready -> spec-review";
-    }
-    if (config.approvalPolicy.planApproval.required) {
-      return "agent-ready -> plan-review";
-    }
-    return "agent-ready -> agent-done";
-  }
-
-  return `${state.kind} -> no-issue`;
-}
-
-function hasBlockedSavedWorkspaceState(
-  state: Awaited<ReturnType<typeof readRunState>>,
-): boolean {
-  return (
-    state?.status === "blocked" &&
-    (state.branch !== undefined || state.worktreePath !== undefined)
-  );
-}
-
-function lifecycleLabels(
-  config: Pick<AgentIssueConfig, "readyLabel" | "triagePolicy">,
-): {
-  ready: string;
-  inProgress: string;
-  done: string;
-  needsInfo: string;
-} {
-  const triagePolicy = config.triagePolicy ?? DEFAULT_TRIAGE_POLICY;
-
-  return {
-    ready: config.triagePolicy?.labels.ready ?? config.readyLabel,
-    inProgress: triagePolicy.labels.inProgress,
-    done: triagePolicy.labels.done,
-    needsInfo: triagePolicy.labels.needsInfo,
-  };
-}
-
-const RESUME_ONLY_SIDE_EFFECT_CHECKPOINTS = new Set<
-  keyof AgentIssueRunCheckpoints
->([
-  "claimed",
-  "startedCommentPosted",
-  "readyLabelRestored",
-  "specReadyCommentPosted",
-  "planReadyCommentPosted",
-  "worktreeReady",
-  "implementationCompleted",
-  "visualEvidenceValidated",
-  "handoffCommentPosted",
-  "doneLabelEnsured",
-  "doneLabelApplied",
-]);
-
-function effectiveCheckpoints(
-  checkpoints: AgentIssueRunCheckpoints | undefined,
-  resumable: boolean,
-): AgentIssueRunCheckpoints | undefined {
-  if (resumable || !checkpoints) return checkpoints;
-
-  const filtered = Object.fromEntries(
-    Object.entries(checkpoints).filter(
-      ([checkpoint]) =>
-        !RESUME_ONLY_SIDE_EFFECT_CHECKPOINTS.has(
-          checkpoint as keyof AgentIssueRunCheckpoints,
-        ),
-    ),
-  ) as AgentIssueRunCheckpoints;
-
-  return Object.keys(filtered).length > 0 ? filtered : undefined;
-}
-
-function startedComment(issue: IssueSummary): string {
-  return `Automation started for issue #${issue.number}.
-
-The issue has been claimed for plan and implementation orchestration.`;
-}
-
-function handoffComment(
-  planPath: string,
-  result: Extract<AgentIssuePiResult, { status: "pr-created" | "merged" }>,
-  baseBranch: string,
-): string {
-  const lines = [
-    `Automation handoff ready.`,
-    ``,
-    `- Plan: \`${planPath}\``,
-    `- Branch: \`${result.branch}\``,
-  ];
-
-  if (result.status === "pr-created") {
-    lines.push(`- PR: ${result.prUrl}`);
-  } else {
-    lines.push(`- Merged to \`${baseBranch}\`: ${result.mergeCommit}`);
-  }
-
-  if (result.landingDecision) {
-    lines.push(`- Landing decision: ${result.landingDecision}`);
-  }
-  if (result.reviewSummary) lines.push(`- Review: ${result.reviewSummary}`);
-  if (result.validation.length > 0) {
-    lines.push(`- Validation:`);
-    lines.push(...result.validation.map((entry) => `  - ${entry}`));
-  }
-
-  return lines.join("\n");
-}
-
-function questionText(question: AgentIssueBlockerQuestion): string {
-  return typeof question === "string"
-    ? `- ${question}`
-    : `- ${question.question}${question.recommendedAnswer ? `\n  Recommended: ${question.recommendedAnswer}` : ""}`;
-}
-
-function blockerComment(result: AgentIssueBlockedResult): string {
-  return [
-    `Automation blocked and needs more information.`,
-    ``,
-    result.reason,
-    ...(result.questions.length > 0
-      ? ["", "Questions:", ...result.questions.map(questionText)]
-      : []),
-  ].join("\n");
-}
-
-function errorMessage(error: unknown): string {
-  return error instanceof Error ? error.message : String(error);
-}
-
-function rejectionMessage(reason: IssueSelectionRejection["reason"]): string {
-  if (reason === "blocking-labels") return "blocking labels";
-  if (reason === "non-open-state") return "non-open state";
-  if (reason === "waiting-spec-approval") return "waiting for spec approval";
-  if (reason === "waiting-plan-approval") return "waiting for plan approval";
-  return "no actionable workflow state";
-}
-
-async function emitSelectionDiagnostics(
-  rejections: IssueSelectionRejection[],
-  options: RunOneIssueOptions,
-): Promise<void> {
-  for (const rejection of rejections) {
-    await progress(
-      options,
-      "debug",
-      "select",
-      `skipped #${rejection.issueNumber}: ${rejectionMessage(rejection.reason)}`,
-      { issueNumber: rejection.issueNumber, data: rejection },
-    );
-  }
-}
-
-class AgentIssueSafetyError extends Error {
-  readonly name = "AgentIssueSafetyError";
-}
-
-function stringArray(value: unknown): string[] | undefined {
-  if (!Array.isArray(value)) return undefined;
-  const entries = value.filter(
-    (entry): entry is string => typeof entry === "string",
-  );
-  return entries.length === value.length ? entries : undefined;
-}
-
-function visualEvidenceArray(
-  value: unknown,
-): AgentIssueVisualEvidence[] | undefined {
-  if (!Array.isArray(value)) return undefined;
-  const entries = value.flatMap((entry): AgentIssueVisualEvidence[] => {
-    if (!entry || typeof entry !== "object") return [];
-    const record = entry as Record<string, unknown>;
-    if (typeof record.screenshotPath !== "string") return [];
-    return [
-      {
-        screenshotPath: record.screenshotPath,
-        caption:
-          typeof record.caption === "string" ? record.caption : undefined,
-        referencePaths: stringArray(record.referencePaths),
-        url: typeof record.url === "string" ? record.url : undefined,
-      },
-    ];
-  });
-  return entries.length > 0 ? entries : undefined;
-}
-
-function successfulImplementationFromState(
-  state:
-    | {
-        implementationStatus?: "pr-created" | "merged";
-        branch?: string;
-        prUrl?: string;
-        mergeCommit?: string;
-        commits?: unknown;
-        validation?: unknown;
-        reviewSummary?: unknown;
-        landingDecision?: unknown;
-        visualEvidence?: unknown;
-      }
-    | undefined,
-):
-  | Extract<AgentIssuePiResult, { status: "pr-created" | "merged" }>
-  | undefined {
-  if (!state?.implementationStatus || !state.branch) return undefined;
-
-  const commits = stringArray(state.commits);
-  const validation = stringArray(state.validation);
-  if (!commits || !validation) return undefined;
-  const reviewSummary =
-    typeof state.reviewSummary === "string" ? state.reviewSummary : undefined;
-  const landingDecision =
-    typeof state.landingDecision === "string"
-      ? state.landingDecision
-      : undefined;
-  const visualEvidence = visualEvidenceArray(state.visualEvidence);
-
-  if (
-    state.implementationStatus === "pr-created" &&
-    typeof state.prUrl === "string"
-  ) {
-    return {
-      status: "pr-created",
-      prUrl: state.prUrl,
-      branch: state.branch,
-      commits,
-      validation,
-      reviewSummary,
-      landingDecision,
-      visualEvidence,
-    };
-  }
-
-  if (
-    state.implementationStatus === "merged" &&
-    typeof state.mergeCommit === "string"
-  ) {
-    return {
-      status: "merged",
-      branch: state.branch,
-      mergeCommit: state.mergeCommit,
-      commits,
-      validation,
-      reviewSummary,
-      landingDecision,
-    };
-  }
-
-  return undefined;
-}
-
-function assertDirectLandAllowed(
-  result: Extract<AgentIssuePiResult, { status: "pr-created" | "merged" }>,
-  config: Pick<AgentIssueConfig, "allowDirectLand" | "skills">,
-  source: string,
-): void {
-  if (result.status !== "merged") return;
-
-  if (!config.allowDirectLand) {
-    throw new AgentIssueSafetyError(
-      `${source} returned merged while git.allowDirectLand is false`,
-    );
-  }
-
-  if (!config.skills.landing) {
-    throw new AgentIssueSafetyError(
-      `${source} returned merged but direct landing requires git.allowDirectLand=true and configured skills.landing`,
-    );
-  }
-}
-
-async function selectResumableIssue(
-  issues: IssueSummary[],
-  config: AgentIssueConfig,
-): Promise<{ issue: IssueSummary; resumed: boolean } | undefined> {
-  const { inProgress, ready } = lifecycleLabels(config);
-  const shouldResume = config.execute && !config.dryRun;
-  const resumable: IssueSummary[] = [];
-  if (shouldResume) {
-    for (const issue of issues) {
-      if (!issue.labels.includes(inProgress)) continue;
-      const state = await readRunState(config.runStateDir, issue.number);
-      if (state && isResumableRunState(state)) resumable.push(issue);
-    }
-  }
-
-  if (resumable.length > 1) {
-    throw new Error(
-      `Multiple resumable ${inProgress} automation runs found: ${resumable.map((issue) => `#${issue.number}`).join(", ")}`,
-    );
-  }
-
-  if (config.issueNumber !== undefined) {
-    const resumableSelected = resumable.find(
-      (issue) => issue.number === config.issueNumber,
-    );
-    if (resumableSelected) {
-      return { issue: resumableSelected, resumed: true };
-    }
-    if (shouldResume) {
-      const explicitIssue = issues.find(
-        (candidate) =>
-          candidate.number === config.issueNumber && candidate.state === "open",
-      );
-      const explicitState = explicitIssue
-        ? await readRunState(config.runStateDir, explicitIssue.number)
-        : undefined;
-      if (explicitIssue && hasBlockedSavedWorkspaceState(explicitState)) {
-        if (
-          resumable.length === 1 &&
-          resumable[0]?.number !== explicitIssue.number
-        ) {
-          throw new Error(
-            `Resumable ${inProgress} automation run #${resumable[0]?.number} exists; resume it before processing #${explicitIssue.number}`,
-          );
-        }
-        return { issue: explicitIssue, resumed: true };
-      }
-    }
-    const selected = selectIssue(issues, {
-      issueNumber: config.issueNumber,
-      readyLabel: ready,
-      triagePolicy: config.triagePolicy,
-      approvalPolicy: config.approvalPolicy,
-    });
-    if (!selected) return undefined;
-    if (resumable.length === 1 && resumable[0]?.number !== selected.number) {
-      throw new Error(
-        `Resumable ${inProgress} automation run #${resumable[0]?.number} exists; resume it before processing #${selected.number}`,
-      );
-    }
-    return {
-      issue: selected,
-      resumed: resumable[0]?.number === selected.number,
-    };
-  }
-
-  if (resumable.length === 1) {
-    return { issue: resumable[0], resumed: true };
-  }
-
-  const diagnostics = selectIssueWithDiagnostics(issues, {
-    issueNumber: config.issueNumber,
-    readyLabel: ready,
-    triagePolicy: config.triagePolicy,
-    approvalPolicy: config.approvalPolicy,
-  });
-  return diagnostics.issue
-    ? { issue: diagnostics.issue, resumed: false }
-    : undefined;
-}
-
-function mergeIssueLists(
-  primary: IssueSummary[],
-  secondary: IssueSummary[],
-): IssueSummary[] {
-  const issues = new Map<number, IssueSummary>();
-  for (const issue of secondary) issues.set(issue.number, issue);
-  for (const issue of primary) issues.set(issue.number, issue);
-  return [...issues.values()];
-}
-
-async function loadSelectionIssues(
-  host: IssueHostProvider,
-  config: AgentIssueConfig,
-  options: RunOneIssueOptions,
-): Promise<IssueSummary[]> {
-  if (config.issueNumber === undefined) {
-    await progress(options, "info", "select", "listing open issues");
-    return host.listOpenIssues();
-  }
-
-  await progress(
-    options,
-    "info",
-    "select",
-    `loading issue #${config.issueNumber}`,
-    { issueNumber: config.issueNumber },
-  );
-  const requestedIssues = [await host.viewIssue(config.issueNumber)];
-  const shouldResume = config.execute && !config.dryRun;
-  if (!shouldResume) return requestedIssues;
-
-  await progress(options, "info", "select", "listing open issues");
-  const openIssues = await host.listOpenIssues();
-  return mergeIssueLists(requestedIssues, openIssues);
-}
-
-function unexpectedFailureComment(
-  reason: string,
-  inProgressLabel: string,
-): string {
-  return [
-    `Automation failed unexpectedly and remains ${inProgressLabel}.`,
-    ``,
-    reason,
-    ``,
-    `A human should inspect the run logs before re-running or relabeling this issue.`,
-  ].join("\n");
-}
-
-function unexpectedFailureCommentKey(
-  status: "claimed" | "planning" | "implementing",
-): string {
-  return `unexpected-failure:${status}`;
-}
-
-async function unexpectedFailure(
-  host: IssueHostProvider,
-  config: AgentIssueConfig,
-  issue: IssueSummary,
-  checkpoints: AgentIssueRunCheckpoints,
-  details: {
-    specPath?: string;
-    specCommit?: string;
-    planPath?: string;
-    planCommit?: string;
-    branch?: string;
-    worktreePath?: string;
-  },
-  timestamp: string,
-  error: unknown,
-  options: RunOneIssueOptions,
-): Promise<AgentIssuePipelineResult> {
-  const reason = errorMessage(error);
-  const status =
-    details.branch ||
-    details.worktreePath ||
-    checkpoints.worktreeReady ||
-    checkpoints.implementationCompleted
-      ? "implementing"
-      : details.specPath ||
-          details.specCommit ||
-          checkpoints.specPathResolved ||
-          checkpoints.specCreated ||
-          checkpoints.specReadyCommentPosted ||
-          details.planPath ||
-          details.planCommit ||
-          checkpoints.planPathResolved ||
-          checkpoints.planCreated ||
-          checkpoints.planReadyCommentPosted ||
-          checkpoints.readyLabelRestored
-        ? "planning"
-        : "claimed";
-  await progress(options, "error", "blocked", `blocked: ${reason}`, {
-    issueNumber: issue.number,
-  });
-  await writeRunState(
-    config.runStateDir,
-    {
-      issueNumber: issue.number,
-      title: issue.title,
-      status,
-      specPath: details.specPath,
-      specCommit: details.specCommit,
-      planPath: details.planPath,
-      planCommit: details.planCommit,
-      branch: details.branch,
-      worktreePath: details.worktreePath,
-      lastError: reason,
-    },
-    timestamp,
-  );
-  const state = await readRunState(config.runStateDir, issue.number);
-  const failureCommentKey = unexpectedFailureCommentKey(status);
-  if (!state?.failureCommentKeys?.includes(failureCommentKey)) {
-    const { inProgress } = lifecycleLabels(config);
-    const commented = await host
-      .commentIssue(issue.number, unexpectedFailureComment(reason, inProgress))
-      .then(() => true)
-      .catch(() => false);
-    if (commented) {
-      await writeRunState(
-        config.runStateDir,
-        {
-          issueNumber: issue.number,
-          status,
-          specPath: details.specPath,
-          specCommit: details.specCommit,
-          planPath: details.planPath,
-          planCommit: details.planCommit,
-          branch: details.branch,
-          worktreePath: details.worktreePath,
-          failureCommentKeys: [failureCommentKey],
-        },
-        timestamp,
-      );
-    }
-  }
-
-  await emitSimpleStep(options, issue.number, "final result blocked");
-
-  return withLogPath(
-    {
-      status: "blocked",
-      reason,
-      questions: [],
-      commits: [],
-      validation: [],
-      ...details,
-      issue,
-    },
-    options,
-  );
-}
-
-async function blockIssue(
-  host: IssueHostProvider,
-  config: AgentIssueConfig,
-  issue: IssueSummary,
-  labels: string[],
-  result: AgentIssueBlockedResult,
-  details: {
-    specPath?: string;
-    specCommit?: string;
-    planPath?: string;
-    planCommit?: string;
-    branch?: string;
-    worktreePath?: string;
-  },
-  timestamp: string,
-  options: RunOneIssueOptions,
-): Promise<AgentIssuePipelineResult> {
-  const { inProgress, needsInfo } = lifecycleLabels(config);
-  await progress(options, "error", "blocked", `blocked: ${result.reason}`, {
-    issueNumber: issue.number,
-  });
-  const blockedLabels = nextLabels(labels, [inProgress], [needsInfo]);
-  await ensureAutomationLabel(host, config, needsInfo);
-  await host.applyLabels(planLabelChange(issue.number, labels, blockedLabels));
-  await writeRunState(
-    config.runStateDir,
-    {
-      issueNumber: issue.number,
-      title: issue.title,
-      status: "blocked",
-      specPath: details.specPath,
-      specCommit: details.specCommit,
-      planPath: details.planPath,
-      planCommit: details.planCommit,
-      branch: details.branch,
-      worktreePath: details.worktreePath,
-      lastError: result.reason,
-      commits: result.commits,
-      validation: result.validation,
-      blockerQuestions: result.questions,
-    },
-    timestamp,
-  );
-  await host
-    .commentIssue(issue.number, blockerComment(result))
-    .catch(() => undefined);
-
-  await emitSimpleStep(options, issue.number, "final result blocked");
-
-  return withLogPath({ ...result, ...details, issue }, options);
-}
 
 export async function runOneIssue(
   runner: CommandRunner,
